set var decisivas

+++ codigo/carrinha.cpp
@@ -37,6 +37,6 @@ int Carrinha::removerEncomenda(Encomenda *encomenda) {
     return 1;
 }
 
-void Carrinha::setVarDecisiva(int peso_peso, int peso_vol) {
+void Carrinha::setVarDecisiva(double peso_peso, double peso_vol) {
     var_decisiva = peso_peso*peso + peso_vol*vol;
 }
