+++ codigo/encomenda.h
@@ -68,7 +68,7 @@ public:
      * @param peso_peso percentagem do peso
      * @param peso_vol percentagem do volume
      */
-    void setVarDecisiva(int peso_peso, int peso_vol);
+    void setVarDecisiva(double peso_peso, double peso_vol);
 
     /**
      * Retorna variavel decisiva
